ap_uint<N> = uint32_t

+++ data/ap_int.h
@@ -2,6 +2,6 @@
 #define RTCORE_HLS_AP_INT_H
 
 template<int N>
-using ap_uint = uintmax_t;
+using ap_uint = uint32_t;
 
 #endif //RTCORE_HLS_AP_INT_H
